+++ src/owned_slice/drain.rs
@@ -11,6 +11,8 @@ use core::{
     ptr::{self, NonNull},
 };
 
+use super::OwnedSlice;
+
 /// A draining iterator for owned slices.
 ///
 /// This struct is created by the `drain` method on
@@ -252,3 +254,71 @@ impl<T> ExactSizeIterator for Drain<'_, T> {
 unsafe impl<T> core::iter::TrustedLen for Drain<'_, T> {}
 
 impl<T> FusedIterator for Drain<'_, T> {}
+
+unsafe impl<T> OwnedSlice for Drain<'_, T> {
+    type Item = T;
+
+    fn owned_slice_ptr(&self) -> NonNull<[Self::Item]> {
+        self.iter.owned_slice_ptr()
+    }
+
+    fn take_owned_slice(&mut self) {
+        self.iter.take_owned_slice();
+    }
+}
+
+#[cfg(all(test, feature = "std"))]
+mod tests {
+    use crate::{tests::TestWrap, Bump};
+
+    #[test]
+    fn owned_slice() {
+        let bump: Bump = Bump::new();
+        let slice = bump.alloc_iter((0..5).map(|v| v.to_string()).map(TestWrap));
+        assert_eq!(
+            slice.iter().map(|v| v.0.clone()).collect::<Vec<_>>(),
+            &["0", "1", "2", "3", "4"]
+        );
+
+        for start in 0..slice.len() {
+            for end in start..slice.len() {
+                TestWrap::expect().drops(5).clones(5).run(|| {
+                    let mut slice_clone = bump.alloc_slice_clone(&slice);
+
+                    let mut vec = bump.alloc_fixed_vec(10);
+                    vec.append(slice_clone.drain(start..end));
+
+                    assert_eq!(vec, slice[start..end]);
+
+                    assert_eq!(
+                        TestWrap::peel_slice(slice_clone.as_slice()),
+                        TestWrap::peel_slice(&slice[..start])
+                            .iter()
+                            .chain(TestWrap::peel_slice(&slice[end..]))
+                            .cloned()
+                            .collect::<Vec<_>>()
+                    );
+                });
+            }
+        }
+    }
+
+    #[test]
+    fn owned_slice_zst() {
+        let bump: Bump = Bump::new();
+        let slice = bump.alloc_iter((0..5).map(|v| v.to_string()).map(TestWrap));
+
+        for start in 0..slice.len() {
+            for end in start..slice.len() {
+                TestWrap::expect().drops(5).clones(5).run(|| {
+                    let mut slice_clone = bump.alloc_slice_clone(&slice);
+                    let mut vec = bump.alloc_fixed_vec(10);
+                    vec.append(slice_clone.drain(start..end));
+
+                    assert_eq!(vec.len(), end - start);
+                    assert_eq!(slice_clone.len(), slice[..start].len() + slice[end..].len());
+                });
+            }
+        }
+    }
+}